refactor(DAO-2135): move deposit modal components into modals/deposit/ subfolder
Move BtcDepositModal, DepositSteps, DepositAmountStep, and
DepositReviewStep (plus tests and stories) into modals/deposit/.
Add barrel index.ts.
