+++ src/app/btc-vault/components/modals/deposit/BtcDepositModal.tsx
@@ -9,10 +9,10 @@ import { ProgressBar } from '@/components/ProgressBarNew'
 import { Header } from '@/components/Typography'
 import { VAULT_SHARE_DECIMALS } from '@/lib/constants'
 
-import { useUserPosition } from '../hooks/useUserPosition'
-import { useVaultMetrics } from '../hooks/useVaultMetrics'
-import { BTC_VAULT_DEPOSIT_FEE, DEPOSIT_STEP_PROGRESS } from '../services/constants'
-import type { DepositRequestParams } from '../services/types'
+import { useUserPosition } from '../../../hooks/useUserPosition'
+import { useVaultMetrics } from '../../../hooks/useVaultMetrics'
+import { BTC_VAULT_DEPOSIT_FEE, DEPOSIT_STEP_PROGRESS } from '../../../services/constants'
+import type { DepositRequestParams } from '../../../services/types'
 import { DepositAmountStep } from './DepositAmountStep'
 import { DepositReviewStep } from './DepositReviewStep'
 import { DepositSteps } from './DepositSteps'
