+++ src/app/btc-vault/components/modals/deposit/index.ts
@@ -1,0 +1,1 @@
+export { BtcDepositModal } from './BtcDepositModal'
